requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0
reportlab==4.0.7
Werkzeug==3.0.1
gunicorn==21.2.0
